♻️refactor: plantService로 클래스 네임 변경
